fixed signed-bitwise issue

+++ src/exhook.c
@@ -5,12 +5,10 @@ static int get_jump_size(void *src, void *dst) {
     int jump_size = 0;
     int64_t gap = (int64_t)dst - (int64_t)src;
 #ifdef __aarch64__
-    if (gap <= 0x8000000 - 1 && gap >= -0x8000000)
-        jump_size = 4;
+    if (gap <= 0x8000000 - 1 && gap >= -0x8000000) jump_size = 4;
     else {
-        gap = ((int64_t)dst >> 12) - ((int64_t)src >> 12);
-        if (gap <= 0x100000 - 1 && gap >= -0x100000)
-            jump_size = 12;
+        gap = (int64_t)((uint64_t)dst >> 12) - (int64_t)((uint64_t)src >> 12);
+        if (gap <= 0x100000 - 1 && gap >= -0x100000) jump_size = 12;
         else {
             jump_size = 4;
             uint64_t imm64 = (uint64_t)dst;
@@ -19,10 +17,8 @@ static int get_jump_size(void *src, void *dst) {
         }
     }
 #elif __x86_64__
-    if (gap <= INT32_MAX && gap >= INT32_MIN)
-        jump_size = 5;
-    else
-        jump_size = 14;
+    if (gap <= INT32_MAX && gap >= INT32_MIN) jump_size = 5;
+    else jump_size = 14;
 #endif
     return jump_size;
 }
